refactor: refactor Model class

+++ include/simple_3d_viewer/linear_algebra/Vertex.hpp
@@ -6,17 +6,19 @@
 
 namespace Simple3D
 {
+
 struct Vertex
 {
   glm::vec3 position;
   glm::vec3 normal;
   glm::vec3 color;
 
-  static constexpr size_t maxUVChannels = 8;
+  static constexpr unsigned int maxUVChannels = 8;
   std::array<glm::vec2, maxUVChannels> texCoords;
 };
 static_assert(sizeof(glm::vec3) == 3 * sizeof(float));
 static_assert(
     sizeof(std::array<glm::vec2, Vertex::maxUVChannels>) ==
     sizeof(glm::vec2) * Vertex::maxUVChannels);
+
 }  // namespace Simple3D
